Module re_remote_store_types::v0::_v0

Modules

column_selector

Nested message and enum types in ColumnSelector.

storage_node_client

Generated client implementations.

storage_node_server

Generated server implementations.

Structs

ColumnSelection

ColumnSelector

Component

Component describes semantic data that can be used by any number of rerun’s archetypes.

ComponentColumnSelector

Used to select a component based on its EntityPath and Component name.

ComponentsSet

EntityPath

The unique identifier of an entity, e.g. camera/3/points See <https://www.rerun.io/docs/concepts/entity-path> for more on entity paths.

GetRecordingMetadataRequest

GetRecordingMetadataResponse

IndexColumnSelector

IndexRange

IndexValues

ListRecordingsRequest

ListRecordingsResponse

ObjectStorage

Query

QueryRequest

QueryResponse

RecordingId

unique recording identifier. At this point in time it is the same id as the ChunkStore’s StoreId

RecordingInfo

RecordingMetadata

RegisterRecordingsRequest

RegisterRecordingsResponse

SampledIndexValues

Schema

arrow IPC serialized schema

TimeColumnSelector

Used to telect a time column.

TimeInt

A 64-bit number describing either nanoseconds, sequence numbers or fully static data.

TimeMetadata

TimeRange

A time range between start and end time points. Each 64 bit number can represent different time point data depending on the timeline it is associated with. Time range is inclusive for both start and end time points.

Timeline

A recording can have multiple timelines, each is identified by a name, for example log_tick, log_time, etc.

ViewContents

ViewContentsPart

Enums

EncoderVersion

RecordingType

SparseFillStrategy

Specifies how null values should be filled in the returned dataframe.